working on compelxity
